refactor: moved expand logic to tool component
